Update patch set 1
Patch Set 1:

I certainly apologize for any breakage. However, I believed that the creation of this repo was delayed by DiRT exercises and it certainly shouldn't have been added to the manifest yet.

You can just remove it from the manifest. As you've noticed, it doesn't interoperate with OpenSSL in the same build!

Patch-set: 1