Update patch set 1
Patch Set 1:

Okay, thought it might have been added so that we could be gradually moving projects over and didn't want to just outright remove it without knowing first (since it's harder to change the manifest than it is a makefile). I'll revert this and upload the change for the manifest.

Patch-set: 1