Update patch set 1
Patch Set 1: Code-Review+1

LGTM as is, but *warning*: there's no ABI compatibility with BoringSSL. Not only is it based on OpenSSL 1.0.2, which has a different ABI than 1.0.1, but we don't maintain ABI compatibility within BoringSSL at all. Even the old patches to OpenSSL would break ABI compatibility.

What are these device binaries? Are they built from the Android tree headers and what's the procedure for that?

Patch-set: 1
Label: Code-Review=+1