Update patch set 1
Patch Set 1:

> Oh, also benchmarks for stuff that has changed (ChaCha20/Poly1305
 > on aarch64, armv7a) when you get the hardware would be welcome!

Oh, that I can get you now. Caveat: this is built with whatever the standalone CMake build does. I don't know if AOSP produces a build of the bssl binary I can toy with.
https://boringssl.googlesource.com/boringssl/+/35be6880788d90bc2a93de5295391b001a8a6a46

(The 6P is a personal device, so I'm happy pushing a random benchmark binary on it. I just didn't want to unlock it and flash a completely new system image. That I'd rather do on the work device when it arrives.)

As a preview, we should also be getting this improvement later:
https://boringssl.googlesource.com/boringssl/+/6f0c4db90e47437ed87a2d385c7797e692a2cf65

I had to revert that because upstream's Poly1305 stuff is broken. As of recently, they've fixed the latest bug, but I'm most of the way through auditing the AVX code in the x86_64 version now, so I think I'll finish that up first and otherwise check that the code the roughly analogous in each of the others. (The most recent bug was present in every architecture's version assembly.)

Patch-set: 1