Update patch set 2
Patch Set 2:

> Not sure what happened to this CL, but there's a way to emit #pragmas under a macro in MSVC (__pragma), so I can guard all our silly warning pragmas with that which should hopefully silence this and we won't need -Wno-unknown-pragmas.

Oh, fun! In fact this does not get rid of all of the unknown pragma warnings, but the others are actually real problems. It seems the thread-local destructors aren't getting hooked up in the MinGW build. So it's probably a good thing we're not just blindly silencing them.

Patch-set: 2