Remove unnecessary libssl_static-host target
This target was used only from external/conscrypt
libconscrypt_openjdk_jni.so which has been switched to the libssl
target instead.

This target is unnecessary because the build system is these days
smart enough to disable sanitization of all dependent libraries of
a target for which sanitization is explicitly disabled, such as the
libconscrypt_openjdk_jni target.

Test: Commit has no effect on the contents of
      libconscrypt_openjdk_jni.so, with and witout ASAN enabled:
      $ make SANITIZE_HOST=address libconscrypt_openjdk_jni
      $ make libconscrypt_openjdk_jni
Bug: 26160319
Change-Id: Ia3375e1e2de174539d603022306855d2ba1761c2