Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I'm like 50% sure the relevant test failure is due to BoringSSL not keeping around PrivateKeyInfo:
> > 
> > https://boringssl.googlesource.com/boringssl/+/02084ea3988446273558881a7e728b04d7338350%5E%21/
> 
> Looking at this now.

It doesn't look like the cause is dropping the attributes. The line that is failing is an assert using OpenSSLECPrivateKey.equals, but the implementation of that only compares the curve and key, not the random attributes on it. (Conscrypt, like every other consumer I've seen, does not retain the attributes to begin with.) I'm not sure this change actually affects code used by that test... it's about the weird PKCS#8 encryption stuff, not normal PrivateKeyInfo serialization.

Still poking to see what the actual cause might be.

Patch-set: 1
Reviewer: Gerrit User 1031048 <1031048@85c56323-6fa9-3386-8a01-6480fb634889>