Update patch set 1
Patch Set 1:

> CTS Run: https://sponge.corp.google.com/target?id=3228aa26-e715-4619-bc00-9bc13e70f468&target=x86+CtsLibcoreTestCases&searchFor=&show=FAILED&sortBy=STATUS

The failure in SSLSocketTest looks like a flake (we've been seeing similar failures pop up elsewhere as well), so a rerun will probably clear that.

The KeyPairGeneratorTest failures look like actual problems, though.  In both of those tests, the failure is being caused by a call to EVP_PKEY_cmp with the same EVP_PKEY* as both arguments that returns -2, with an error of INCOMPATIBLE_OBJECTS at ec.c:713 (which is in EC_POINT_cmp).  Can you look into that?

Patch-set: 1