Update patch set 2
Patch Set 2:

> > CTS run: https://sponge.corp.google.com/target?id=9b7ebcbb-dec7-463a-9647-c44f24780a30&target=x86+CtsLibcoreTestCases&searchFor=&show=FAILED&sortBy=STATUS
 > >
 > > One possibly relevant failure in
 > > SSLSocketTest.test_SSLSocket_setUseClientMode. Investigating...
 > 
 > This is likely spurious: test_SSLSocket_setUseClientMode(false,
 > false) --
 > which tries to construct a server->server socket -- fails with a
 > SSLHandshakeException as opposed to a SocketTimeoutException. Still
 > looking for why though.

Hrm, yeah, I'd also be interested in why that's happening.  As I understand it, both sockets should wait forever for the other side to send a client hello, so timeout makes sense.  I'm not sure what could cause something else to happen, unless timeout perhaps is being communicated in a different way than before and we're not seeing it as a timeout in Conscrypt?

Patch-set: 2