Update patch set 1
Patch Set 1:

> Also @agl -- you want I should make these tests run continuously after this CL is merged?

Sadly not: we can't publish the challenges and we can only sort-of check whether we think we passed because NIST keeps the checking tool private. This is a tool for manually running the tests in cooperation with a NIST-approved lab.

Patch-set: 1
Reviewer: Gerrit User 1019036 <1019036@85c56323-6fa9-3386-8a01-6480fb634889>