Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> +davidben.

Thanks! Here's the upstream one:
https://boringssl-review.googlesource.com/c/boringssl/+/27666

> (i assume they'll want to take this change upstream first, and Android will take it next time this project is synced with upstream.)

Yeah, that's usually the preference. Otherwise we might accidentally lose changes. Is this just cleanup or is it blocking something? If it's blocking something, we can land it in both and coordinate with varomodt@ to make sure it doesn't get lost.

Patch-set: 1
Reviewer: Gerrit User 1031048 <1031048@85c56323-6fa9-3386-8a01-6480fb634889>