Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Can you explain what changes in the recovery variant of the library?

From the perspective of runtime behavior, nothing is changed. The change is only at build-time. Before the platform/build/soong CL in the this topic, you build a library only once for the system partition and statically link to it if the library is required by an executable that is to be installed to the recovery partition.

Now, the build system (soong) officially supports the recovery partition and it has not been allowing cross-partition dependencies for all types of modules including static libs. So, if something (header libs, static libs, shared libs) has to be used for non-system partitions as well as the system partition, it has to be marked as vendor_available (for vendor partition) and recovery_available (for recovery partition). Then the module is built twice: once for the system and once for the non-system (recovery) partition. The executable in recovery partition will be linked against the latter variant of the library.

The reason this has worked without recovery_available is because the modules (e.g. toybox) using this static lib has been built in build/make where the cross-partition dependencies haven't been enforced. Now, toybox (and others bins in the recovery partition) are moving to soong and this becomes mandatory.

This also paves the way for supporting shared libs in the recovery partition.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>