Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Sorry for inappropriate reply. This CL actually does not change anything about the library. Instead it just reflects the current status, that the different variants of this lib can be loaded in a same vendor process. This possibility already exists because this lib has a vendor variant and also is depended on by LLNDK. By marking it explicitly, a build system can track this behavior through the dependency chain.
> 
> I'm not quite following. Could you elaborate on what exactly double-loading here means? (BoringSSL does have unavoidable global state like the PRNG. Depending on the answer, this may not actually be safe.)

It means loading the library at two different places in memory. Global state is fine, but if for instance, functions took process-wide ownership of something from the kernel OR talked to other services in specific ways, it could be problematic.

It's worth noting that this is already happening (running in this configuration), and this build system change is aimed at preventing the spread of these dependencies (and labeling things so that we can even look at preventing this).

Patch-set: 2
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>