Update patch set 2
Patch Set 2:

This CL is not blocked by the others in this topic, but the others in this topic depend on this CL in order to work correctly (plus there is additional work required).

I'm adding this CL to the CL topic so that it's not overlooked, but if we can resolve the build failure then we can also submit this one before the other ones.

I also have a local CL to drop the libssl_static build target (in the hope that it will fix the treehugger failure), but that target is still used by an internal target, which I will not be able to delete until after Colin's CLs from ~ 14 hours ago have merged to the internal master branch, which was not yet the case when I checked an hour ago.

Patch-set: 2
CC: Gerrit User 1103203 <1103203@85c56323-6fa9-3386-8a01-6480fb634889>