Update patch set 1
Patch Set 1: -Autosubmit

> I'm not sure that the .rc will be picked up if it lives here (I'm not even sure if it was picked up in the previous place, although .rc files in apex/ are picked up.

Note that the .rc file here is intended for the vendor partition, so I'm not sure why it wouldn't work here when it was working OK in external/conscrypt.

I'll test it before submitting by patching to a tree without a prebuilt vendor.img and observing that (a) the .rc file gets copied to /vendor/etc/init and that the run of the self test is logged to logcat.  Is that sufficient?

Patch-set: 1
Reviewer: Gerrit User 1334152 <1334152@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=0
Label: Presubmit-Ready=0