Update patch set 1
Patch Set 1: Code-Review+2

Happy to +2, but if this is part of the work to update LLVM then you can maybe remove the "hwaddress: false" in preference?  That was added due to Clang crashing with linker scripts on hwasan builds,
detailed in b/142971559.

Patch-set: 1
Reviewer: Gerrit User 1334152 <1334152@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2