Update patch set 1
Patch Set 1:

Note that libcrypto_static is the non-FIPS version whereas the non-static version is the FIPS version.

Is there a particular reason why you need specifically the static (non-FIPS) version, or can you use the nonstatic (FIPS) version? If so, what do you need it for? (Whether or not it's okay to use non-FIPS version for your use case might need to be subject to review)

Patch-set: 1
Reviewer: Gerrit User 1103203 <1103203@85c56323-6fa9-3386-8a01-6480fb634889>