Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Paul, as discussed in other CL, boringssl is not part of any SDK - conscrypt statically includes it (meaning you don't need this CL)

You may be being confused by the title, should have read module_exports not sdk.

Just because it not part of an SDK (i.e. not something provided by an APEX) does not mean that a prebuilt of it it is not needed. As long as external/boringssl is part of the ART module, Conscrypt uses it and we have to support unbundled builds then we need to export boringssl prebuilts from the ART module.

Patch-set: 1