Update patch set 1
Patch Set 1:

> i pasted in the socket failure:

Ah, sorry. I see it now. Right, that. It shouldn't need a network, but it does want a working v4 or v6 loopback. Does Android's test environment have a story for that sort of thing, or is the expectation that you don't create sockets at all? We can certainly arrange to disable that test on Android if necessary, though I would imagine we're not the only ones making *some* kind of socket.

Patch-set: 1