Update patch set 1
Patch Set 1:

> hmm, i'd have expected loopback to work. i think we use that elsewhere in tests.

Looking at the test some more, it's part of the "connect" BIO we inherited from OpenSSL (there's a lot of goofy OpenSSL compat stuff in here). It expects a hostname string that gets passed to getaddrinfo, but the tests should be passing in "::1" or "127.0.0.1".

Perhaps whatever network-less setup the tests are running in doesn't make that work? But the weird thing is error is coming from a much later line number. So somehow getaddrinfo() has succeeded and even given us back a sockaddr (which I would hope is the loopback...). Then we try to connect() to that address and it fails.

Patch-set: 1