Use min_sdk_version instead of sdk_version
Boringssl modules including libssl and libcrypto have been built with
`sdk_version: "9"` inherited from boringssl_flags. It looks like the
intent of the setting was to make sure that the modules don't use NDK
APIs newer than the SDK level 9, probably for the maximum backwards
compatibility.

However, the setting is considered too strong because the modules should
be able to check the availability of the newer APIs at runtime and use
the APIs when they are available (i.e. when running on 9+ platforms).
By not setting `sdk_version` but setting `min_sdk_version` to 9, the
modules are linked with the latest APIs but any use of post-9 APIs is
guaranteed to trigger a compilation error when not guarded with
`if __builtin_available(android 9,*)`.

Keep using sdk_version has another side effect for Mainline modules. The
build system, when doing unbundled builds, hides the platform variant of
a module if the module is with `sdk_version` in the Make world. This in
turn causes a side effect that the symbol files for the boringssl
modules which are included in the conscrypt APEX are not installed,
because the installation of the symbol files are done by the Make build
system.

Fixing the issue by switching from sdk_version to min_sdk_version.

Bug: 180622230
Test: DIST_DIR=out/dist TARGET_BUILD_APPS="com.google.android.conscrypt"
./vendor/google/build/build_unbundled_mainline_module.sh
Examine the symbols zip file. libcrypto.so, libssl.so are there.

Change-Id: I59aeed971201f64eadb824c2cd46a3e1908f0ff3