Use genrule to pull in bindgen-erated source
Move the Android build to more closely match the upstream CMake
mechanism, by using a genrule to replace the `${INCLUDES}` line
in the Rust source with an include of the bindgen-created source
code.

This change means that the upstream BoringSSL source no longer has a
local modification, making it easier to import new versions.

Moving to source-inclusion means that:
 - there's no longer a need for a rlib dependency on the generated
   source (as its source is included)
 - the shared_libs dependencies on the crypto code need to move to
   the `libbssl_ffi` target, which also means that...
 - the bindgen source now needs a `local_include_dirs` pointer to
   `src/include`.

Also add a rust_test for the bindgen layout tests along the way.

Bug: 177080016
Test: build targets, atest libbssl_sys_raw_test
Change-Id: Iabf1b6ac4ae7d3a391cc038b29f157347fb36f2f