Ensure clients can't block prng_seeder
Use async to rework the server so that it can serve new clients while
old clients are still reading what they've been sent. Add some wiggle room
on exactly when we reseed the DRBG to support the async case - we
enforce reseed every 256 requests but we try to reseed every 128. Use
proper Android logging and always log on exit.

Bug: 243933553
Test: DO NOT SUBMIT
Change-Id: Icafa5f27f01c5c453bd8dd1965fb5b6a4ad27dbb