Add libcrypto_baremetal
BoringSSL is already used in bare-metal environments, eg. in
external/u-boot, but so far hasn't been compiled that way in Soong.
Now that we are building pvmfw in AOSP, add a libcrypto_baremetal
target that will allow us to link it against BoringSSL.

The target disables the Android userspace stack protector, and defines:
 - OPENSSL_NO_THREADS_CORRUPT_MEMORY_AND_LEAK_SECRETS_IF_THREADED
   - no threading support in this environment
 - OPENSSL_SMALL
   - reduce size by removing uncommon functionality
 - DOPENSSL_NO_ASM
   - remove dependency on getauxval() and __system_property_get

Bug: 237372981
Test: m pvmfw_bin
Change-Id: Ie412d2f9178b183162c9be111a3b4496ac6243b0