Revert "Add libcrypto_baremetal"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/28b11a45-abde-4858-a87f-9a95ada2c20b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6656ffd3bfbd54e66bfa9def7a7b53f51a4b42b7