Revert "external/boringssl: Sync to e40d0f8ee1361fbff2927a6806c7..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/42f7749f-db46-4b92-a176-e2eab9e81478). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idfd10e9f64ed2409c7d68e3179d425bcf6fadfa5