Revert "external/boringssl: Sync to e40d0f8ee1361fbff2927a6806c7..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7a5d9f77-dd0e-4480-9dbf-5fcf62cb9d74). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icc59b2573a80a319a6d3dd8aa5251abbc7fadbab