Revert "external/boringssl: Sync to e40d0f8ee1361fbff2927a6806c7..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c69223ef-2ee2-4437-8e55-45af9c2a2104). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I377d20ebd2288a8c48e4a4668b180464fc17393f