Revert "external/boringssl: Sync to e40d0f8ee1361fbff2927a6806c7..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b7696a0b-4548-4039-929b-d68bcde4d1ca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie81a95da57574d8070789396fc14805e2e6b6ba4