Revert "external/boringssl: Sync to e40d0f8ee1361fbff2927a6806c7..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/017f6ad8-2ef3-42ed-884d-b7b98a1b39be). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idee8e861a0573b89431a6ed062dc5612ec665451