Revert "Re-land January BoringSSL. Revert submission 2446033-rev..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2941e71c-e496-4a4e-be19-f2d73febf3f2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I46dc54a14bd4cdd59ac22608b43905c545659fe6