Revert "Re-land January BoringSSL. Revert submission 2446033-rev..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/db75aec1-7dad-4597-9c17-431960cb4ceb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e33cdcf8a687d42a61293f03a6091a777385bbd