Revert "Re-land January BoringSSL. Revert submission 2446033-rev..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ace545ef-b435-42ae-b04d-51ff62b2712c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie7c73414c16dc7930c07deb15bc5965825163934