Revert "Re-land January BoringSSL. Revert submission 2446033-rev..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72df7ba5-10cd-4fba-94f1-7a694aa8e464). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iac3ef96abe6ecf0bc26e629e878f3b2ad8b597c4