libcrypto_baremetal: Require OPENSSL_cpuid_setup
To allow client (baremetal) code to dynamically configure the library
based on the CPU features discovered, disable STATIC_ARMCAP. This will
cause the library to call the internal function OPENSSL_cpuid_setup() to
record the discovered features in OPENSSL_armcap_P (on ARM), which is
used by the implementation to select the best implementations of the
algorithms.

As we also set __TRUSTY__, no backend implementing the function will be
built, producing a (static) library with an undefined symbol. This
allows client code to hook into the library initialization
process by defining the function and configuring OPENSSL_armcap_P.

Bug: 265125189
Test: m pvmfw_bin  # defines OPENSSL_armcap_P in aosp/2517576
Change-Id: Ia75a1f434265a12e46e39ec24d08d70a989b13f3