Revert "libcrypto_baremetal: Don't set OPENSSL_NO_ASM"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ae98094-9204-4de3-a486-48d7c9773d0e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I60bc343fbef6a5e46162057483fbf355a8c61745