Revert "libcrypto_baremetal: Set NO_STATIC_INITIALIZER"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0a25d50c-d930-48ec-8173-44c305798f96). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia1f2d9a3fffbb5799c26101d9e85085187e3e2e7