Revert "libcrypto_baremetal: Set NO_STATIC_INITIALIZER"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5037d5ae-8a5b-454b-9d09-7fc243cb1f0e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I66f9b24e89bb49205394a9e0a303feec2f9fba70