Revert "libcrypto_baremetal: Don't set OPENSSL_NO_ASM"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/201a203e-5871-47cf-a6fe-b3ebea3d1dd6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibf24dbede092b73c43c797775909f192a2d39e22