Revert "libcrypto_baremetal: Don't set OPENSSL_NO_ASM"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/279f89eb-b4e9-4b30-ba1f-7cedda9b172a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I799c289c1a2eaeaf2dcae892b891f172a80dd8b5