Revert "libcrypto_baremetal: Set NO_STATIC_INITIALIZER"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd7940ec-aaf1-45ff-aad3-808d7708d899). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia9a3c5f2914e26840bc38488ff3aaf900629cc1a