Revert "libcrypto_baremetal: Don't set OPENSSL_NO_ASM"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd7940ec-aaf1-45ff-aad3-808d7708d899). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8f5e6ed0ccf151937ec276b793b1687f6858ebaf