Revert "libcrypto_baremetal: Set NO_STATIC_INITIALIZER"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/589d22a8-67e8-4590-82ea-6a68fc964d78). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7f0e438c7492c4b627739accf83a6a02fcd17c15