Revert "libcrypto_baremetal: Set NO_STATIC_INITIALIZER"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ab222795-d30f-4c7f-99e9-41ff0be0feed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieab5a356b193080040b26ab72ea76e3945575e0d