Revert "libcrypto_baremetal: Don't set OPENSSL_NO_ASM"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de841e63-682e-4cd8-bee5-d0b0cb5518aa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I39e750029775c93e71ec55673a2ec906512a122b