Revert "libcrypto_baremetal: Don't set OPENSSL_NO_ASM"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe4e4a9d-2cf6-4960-9df2-45f7c2fc06c8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I93f23d30adc77f28a6b5a849b470b05e6aa899ce