Revert "libcrypto_baremetal: Set NO_STATIC_INITIALIZER"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe4e4a9d-2cf6-4960-9df2-45f7c2fc06c8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic40dbf761794bbeb4371abdaddfddc92e3bcd0c1