Revert "libcrypto_baremetal: Set NO_STATIC_INITIALIZER"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b06c2a55-f6ea-40cb-8d0c-e03a2c4d1149). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I617877e4a1708fbf600407b10b530865f26da4e5