Revert "libcrypto_baremetal: Don't set OPENSSL_NO_ASM"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b06c2a55-f6ea-40cb-8d0c-e03a2c4d1149). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I904d8e1087008dade1072daecfbbee59c9ac316a