Revert "Diasble memtag_globals for boringssl."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e0a096cc-5d6a-4648-a62c-a26102c31283). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icb0e9a6c9e7da5e40145b9fe972edbd75515ed48