Revert "Diasble memtag_globals for boringssl."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fa254277-2a4a-4e07-b6aa-f9051e30f64e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I31f2ece7206ab70f3185af26b6f5d989e81e1c27