Revert "Diasble memtag_globals for boringssl."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e347ad4f-c007-48ec-8652-922c7a519413). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I78c68f86ca22fe9c6bc8bf6dbc4c2fcda0244fef