Revert "Diasble memtag_globals for boringssl."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c32a908-2227-458a-b8ab-f6cd9e2c0d12). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ife2fabd19a36563e1a830522dbabc9771de5cb05