Revert "external/boringssl: Sync to 26ecb2a275ca7444d10899b8a3fe..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a5e9c19c-5321-471d-8ae0-e94bbd1c1ccd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id5e2e79a6a18b4028cea5437f0a30ffa56b62f0f