Don't define __TRUSTY__ in libcrypto_baremetal
This needs to land together with a pvmfw change to rename CRYPTO_sysrand
to getentropy. See b/290898063.

As part of this, to pick up getentropy, I needed to fix a couple of
issues in pvmfw's "baremetal" build:

- libcrypto_baremetal was built for the host due to accidentally
  inheriting a value from libcrypto_defaults.

- libcrypto_baremetal inherited the SDK version of the real libcrypto.
  This caused the getentropy to prototype to be excluded in the SDK
  build. None of this should apply because libcrypto_baremetal isn't
  building for Android proper in the first place, so none of this is
  applicable.

  Rather, this is a symptom of libcrypto_baremetal being defined as
  building for an Android device, and then it gets linked to different
  libraries than usual. It should instead be defined as its own
  toolchain. (See also b/291101350.)

This rearranges the defaults in the build to avoid both of these. As
part of this, I've pulled OPENSSL_NO_ASM into boringssl_flags, where it
should be (BoringSSL's tests care about this define), and split up some
of the baremetal and normal targets.

The real libcrypto's sdk_version of "9" is almost certainly incorrect
and seems vestigial. I tried setting it to "minimum" and got a
mysterious build failure with test_suites_x86_64-trunk_staging_coverage,
so leaving it at "9" for now.

Bug: 291102972
Test: treehugger
Change-Id: I83f0934535884f9864a602a08618851b8925cc43