external/boringssl: Sync to 6738d477b8763abf38f6d1b60b6d799167ce9c8c.
As part of this, remove OPENSSL_STATIC_ARMCAP from Android baremetal.
BoringSSL now knows to query the Arm system register.

This includes the following changes:

https://boringssl.googlesource.com/boringssl/+log/0ffd3658dcdc21a6c56d234cf2a6008487dcfaa7..6738d477b8763abf38f6d1b60b6d799167ce9c8c

* Restore OPENSSL_RAND_TRUSTY
* Remove old asm source lists from Android sources.mk
* Trim some redundant Arm feature detection files
Update-Note: This is a slight change to the OpenBSD build. Previously,
we assumed OpenBSD on 32-bit Arm implies NEON. Now, we pick it up from
the __ARM_NEON define from ACLE, i.e. whether the compiler has been told
that NEON is available. (This comes from -march or other options.) Doing
that is desirable anyway: if NEON is in your baseline, you should tell
your compiler so it can vectorize loops.
* Stop duplicating the OPENSSL_armcap_P declaration
* arm64: Implement baremetal OPENSSL_cpuid_setup
* Switch the generated Soong (Android) build to the new asm source list
* Restore fuzz/session_corpus
* Mark all of bssl::Span as constexpr
* Sync pki to chromium aef592bb1b83f019c8342bbc2c476758ccb6098d
* Reflect OPENSSL_NO_SOCK and OPENSSL_NO_POSIX_IO into headers

Test: treehugger
Bug: 291106677
Change-Id: I31df79442bc557d8dcff96907991718a29ce2ee3