Revert "external/boringssl: Sync to 7ae2b910c13017b63f1a8bd6c8de..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e4feb6fe-d0df-4e0d-9600-1fdf6a0f2ca8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe27737d303bea48c2525b51d35a0a817b000a15