Revert "external/boringssl: Sync to 7ae2b910c13017b63f1a8bd6c8de..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3a3bbc40-0b46-4fae-a356-24aef3e45b45). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I08a6e17f1f90ba574fc77552d7674adfc9faec79