Remove redundant (and now incorrect) OPENSSL_NO_ASM defines in Android.bp
https://boringssl.googlesource.com/boringssl/+/e4acd6cb568214b1c7db4e59ce54ea2e1deae1f5,
imported to Android in aosp/2689386, changed the sources.bp generation so
that sources.bp handled the OPENSSL_NO_ASM define where needed, and also
fixed a long-standing feature gap where we needlessly disabled assembly
optimizations on macOS.

It was meant to be paired with this change when imported to AOSP, but I
forgot to note it in a BoringSSL Update-Note, so this was lost. This adds
the missing change.

Bug: 294399371
Change-Id: I846cb3bf67d465017e31b9879d403ef689191823
Test: treehugger