Limit bcm_object to devices
bcm_object is a cc_object (ld -r) for the FIPS build. However, ld -r
does not work on darwin's linker. This was working on accident before
because cc_object would skip ld -r if there were no longer script and
only a single source to build. Enabling assembly optimizations breaks
this assumption.

This is fragile. Instead, just limit bcm_object to devices and add a
separate static library version for the host. On the host, we use that
instead.

Bug: 294399371
Test: treehugger
Change-Id: If6d538618d8e49f9452b2bc3515aa23d410c3ae8