Revert "external/boringssl: Sync to 7ae2b910c13017b63f1a8bd6c8de..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b2184449-2f13-429a-a1fd-8d54af9fdad8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia88bc020a0eea17f4c396068a1d1807545b8b6f3