Revert "Migrate 8 crates to monorepo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/91ac2ea1-310c-4823-af42-21e2d30ef537). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373666277
Change-Id: Ia0b70ac62e9d8f9b93541aaee9b3d1c5ddb5fec0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com