Revert "Migrate 8 crates to monorepo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9928a64-e7da-4ae0-a413-a2077df34e28). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373654066
Change-Id: I44e3ae6ce04f286f9de25ddd076e275347aa2728
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com