Revert "Migrate 8 crates to monorepo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1412fbf-e2c1-4b03-95f5-8ccc64a89480). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373666970
Change-Id: I8858de72e70f64feaca2501f28feb85ba525cbbe
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com