Revert "Migrate 8 crates to monorepo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b990d7f3-5331-4d31-935c-a55a40c6057c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373653829
Change-Id: Idc5b316d702632bc2dc65eb3a4557535fbf835a6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com