Revert "Add libopenssl to the configinfrastructure apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/af4f2839-6e09-4392-ac83-1fd66f037fdd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia19749da32e8360ef3f6ca540ae1e6f9fb187642
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com